requirements.txt
requests==2.32.4
python-dotenv==1.1.0
spotipy==2.25.1